feat(deps): add @svgr/webpack dependency
This commit adds the `@svgr/webpack` package as a new development dependency.

This loader will be used in the Webpack configuration to transform SVG files into React components. This allows for easier and more efficient handling of SVG assets, such as icons, by importing them directly as components in the application.